add test/tsconfig.json for vscode
This is mostly the same as our normal tsconfig---it includes src+test instead of just src.

I do not think it is being used for anything other than VS code.

With only the /server/tsconfig.json,
- VS Code reports Request.context does not exist (This is an extra property we added to the Express.Request interface)
- but ts-jest says the tests all compile and pass

So it seems that ts-jest understood correctly that Express.Request.context existed, but VS Code did not.